# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: MyCryptoCoin Payment Gateway API
  description: |
    MyCryptoCoin is the Stripe of crypto payments. Accept Bitcoin, Ethereum, USDT,
    and 20+ cryptocurrencies with a single integration. 0.5% per transaction, no
    hidden fees.

    ## Authentication
    All API requests require authentication via either:
    - **API Key**: Pass your key in the `X-API-Key` header
    - **Bearer Token**: Pass a JWT in the `Authorization: Bearer <token>` header

    API keys are prefixed:
    - `mcc_live_` for production
    - `mcc_test_` for sandbox/test mode

    ## Rate Limits
    Default rate limits are returned in response headers:
    - `X-RateLimit-Limit` — Requests allowed per window
    - `X-RateLimit-Remaining` — Requests remaining
    - `X-RateLimit-Reset` — Unix timestamp when the window resets

    ## Pagination
    List endpoints support cursor-based pagination:
    - `page` — Page number (default: 1)
    - `limit` — Items per page (default: 20, max: 100)

    Response includes `pagination` object with `total`, `page`, `limit`, `pages`.

    ## Errors
    All errors follow a consistent format:
    ```json
    {
      "error": {
        "code": "invalid_request",
        "message": "Human-readable description",
        "details": {}
      }
    }
    ```
  version: "1.0.0"
  contact:
    name: MyCryptoCoin Developer Support
    email: developers@mycrypto.co.in
    url: https://mycrypto.co.in/support
  license:
    name: Proprietary
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login or /auth/refresh
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: "API key prefixed with mcc_live_ or mcc_test_"
  headers:
    X-RateLimit-Limit:
      schema:
        type: integer
      description: Maximum requests per rate limit window
    X-RateLimit-Remaining:
      schema:
        type: integer
      description: Remaining requests in the current window
    X-RateLimit-Reset:
      schema:
        type: integer
      description: Unix timestamp when the rate limit window resets
    X-Request-Id:
      schema:
        type: string
        format: uuid
      description: Unique identifier for the request, useful for support
  schemas:
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Machine-readable error code
              example: invalid_request
            message:
              type: string
              description: Human-readable error message
              example: "The 'amount' field is required."
            details:
              type: object
              additionalProperties: true
              description: Additional context about the error
              example:
                field: amount
                reason: required

    # ─── Auth Schemas ─────────────────────────────────────────────
...
    RefreshTokenRequest:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
          example: "mcc_rt_x7k9m2p4q8..."
    RefreshTokenResponse:
      type: object
      properties:
        access_token:
          type: string
          example: "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9..."
        refresh_token:
          type: string
          example: "mcc_rt_n3w7t0k3n..."
        token_type:
          type: string
          example: Bearer
        expires_in:
          type: integer
          example: 86400
    ForgotPasswordRequest:
      type: object
      required:
        - whatsapp_number
      properties:
        whatsapp_number:
          type: string
          example: "+919876543210"
    ForgotPasswordResponse:
      type: object
      properties:
        message:
          type: string
          example: "Password reset OTP sent to WhatsApp."
        otp_sent:
          type: boolean
          example: true
    ResetPasswordRequest:
      type: object
      required:
        - whatsapp_number
        - otp
        - new_password
      properties:
        whatsapp_number:
          type: string
          example: "+919876543210"
        otp:
          type: string
          example: "482916"
        new_password:
          type: string
          format: password
          minLength: 8
          example: "NewSecurePass456"
    ResetPasswordResponse:
      type: object
      properties:
        message:
          type: string
          example: "Password reset successfully."

    # ─── Payment Schemas ──────────────────────────────────────────
    CreatePaymentRequest:
      type: object
      required:
        - amount
        - currency
        - crypto
      properties:
        amount:
          type: number
          format: double
          minimum: 0.01
          description: Payment amount in the specified fiat currency
          example: 99.99
        currency:
          type: string
          minLength: 3
          maxLength: 3
          description: "ISO 4217 fiat currency code"
          enum: [USD, EUR, GBP, INR, AED, SGD, AUD, CAD, JPY]
          example: USD
        crypto:
          type: string
          description: Cryptocurrency to accept
          enum: [BTC, ETH, USDT, USDC, BNB, SOL, MATIC, DOGE, LTC, XRP, ADA, DOT, AVAX, TRX, LINK, UNI, SHIB, APE, ARB, OP]
          example: USDT
        description:
          type: string
          maxLength: 500
          description: Description shown to the customer
          example: "Premium subscription - March 2026"
        metadata:
          type: object
          additionalProperties:
            type: string
          maxProperties: 20
          description: "Up to 20 key-value pairs for your internal reference"
          example:
            order_id: "ORD-12345"
            customer_email: "buyer@example.com"
        callback_url:
          type: string
          format: uri
          description: URL to receive webhook notifications for this payment
          example: "https://yoursite.com/webhooks/crypto"
        redirect_url:
          type: string
          format: uri
          description: URL to redirect the customer after payment
          example: "https://yoursite.com/order/success"
        expiry_minutes:
          type: integer
          minimum: 5
          maximum: 1440
          default: 30
          description: "Minutes before the payment expires (default: 30)"
          example: 30
    Payment:
      type: object
      properties:
        id:
          type: string
          description: Unique payment identifier
          example: "pay_1a2b3c4d5e6f"
        merchant_id:
          type: string
          example: "mch_a1b2c3d4e5f6"
        status:
          type: string
          enum:
            - pending
            - confirming
            - confirmed
            - settled
            - expired
            - failed
            - refunded
          description: |
            Payment lifecycle:
            - `pending` — Awaiting customer payment
            - `confirming` — Transaction detected, waiting for block confirmations
            - `confirmed` — Required confirmations reached
            - `settled` — Funds credited to merchant wallet
            - `expired` — Payment window expired with no transaction
            - `failed` — Transaction failed or underpaid
            - `refunded` — Payment was refunded
          example: confirmed
        amount:
          type: number
          format: double
          description: Requested fiat amount
          example: 99.99
        currency:
          type: string
          example: USD
        crypto:
          type: string
          example: USDT
        crypto_amount:
          type: string
          description: Exact crypto amount to pay (as string for precision)
          example: "99.990000"
        exchange_rate:
          type: string
          description: Exchange rate at time of creation
          example: "1.0000"
        fee_amount:
          type: string
          description: "MyCryptoCoin fee (0.5%)"
          example: "0.499950"
        fee_percent:
          type: string
          example: "0.5"
        net_amount:
          type: string
          description: Amount after fee deduction
          example: "99.490050"
        deposit_address:
          type: string
          description: Unique address generated for this payment
          example: "0x742d35Cc6634C0532925a3b844Bc9e7595f2bD18"
        network:
          type: string
          description: Blockchain network
          example: "ethereum"
        tx_hash:
          type: string
          nullable: true
          description: On-chain transaction hash (null until detected)
          example: "0xabc123def456..."
        confirmations:
          type: integer
          description: Current confirmation count
          example: 12
        required_confirmations:
          type: integer
          description: Confirmations needed for this crypto
          example: 12
        description:
          type: string
          example: "Premium subscription - March 2026"
        metadata:
          type: object
          additionalProperties:
            type: string
          example:
            order_id: "ORD-12345"
            customer_email: "buyer@example.com"
        callback_url:
          type: string
          example: "https://yoursite.com/webhooks/crypto"
        redirect_url:
          type: string
          example: "https://yoursite.com/order/success"
        checkout_url:
          type: string
          description: Hosted checkout page URL for this payment
          example: "https://pay.mycrypto.co.in/pay_1a2b3c4d5e6f"
        qr_code_url:
          type: string
          description: QR code image URL for the deposit address
          example: "https://api.mycrypto.co.in/v1/payments/pay_1a2b3c4d5e6f/qr"
        expires_at:
          type: string
          format: date-time
          example: "2026-03-19T11:00:00Z"
        confirmed_at:
          type: string
          format: date-time
          nullable: true
          example: "2026-03-19T10:45:22Z"
        settled_at:
          type: string
          format: date-time
          nullable: true
          example: "2026-03-19T10:45:30Z"
        created_at:
          type: string
          format: date-time
          example: "2026-03-19T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2026-03-19T10:45:30Z"
    PaymentList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Payment'
        pagination:
          $ref: '#/components/schemas/Pagination'
...
  responses:
    BadRequest:
      description: Invalid request parameters
      headers:
        X-Request-Id:
          $ref: '#/components/headers/X-Request-Id'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: invalid_request
              message: "The 'amount' field must be greater than 0."
              details:
                field: amount
                reason: minimum
    Unauthorized:
      description: Authentication required or token expired
      headers:
        X-Request-Id:
          $ref: '#/components/headers/X-Request-Id'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: authentication_required
              message: "Invalid or expired authentication token."
...
    NotFound:
      description: Resource not found
      headers:
        X-Request-Id:
          $ref: '#/components/headers/X-Request-Id'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: not_found
              message: "Payment pay_1a2b3c4d5e6f not found."
    TooManyRequests:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          $ref: '#/components/headers/X-RateLimit-Limit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/X-RateLimit-Remaining'
        X-RateLimit-Reset:
          $ref: '#/components/headers/X-RateLimit-Reset'
        X-Request-Id:
          $ref: '#/components/headers/X-Request-Id'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: rate_limit_exceeded
              message: "Too many requests. Retry after 30 seconds."
              details:
                retry_after: 30
    InternalError:
      description: Internal server error
      headers:
        X-Request-Id:
          $ref: '#/components/headers/X-Request-Id'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: internal_error
              message: "An unexpected error occurred. Please try again later."
paths:
...
  /auth/refresh:
    post:
      operationId: refreshToken
      tags: [Authentication]
      summary: Refresh access token
      description: |
        Exchange a valid refresh token for a new access token and refresh token pair.
        The old refresh token is invalidated (rotation).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Tokens refreshed
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshTokenResponse'
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error:
                  code: invalid_refresh_token
                  message: "Refresh token is invalid or has been revoked."
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [Authentication]
      summary: Request password reset OTP
      description: Sends a 6-digit OTP to the merchant's WhatsApp number for password reset.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        '200':
          description: OTP sent
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForgotPasswordResponse'
        '404':
          description: WhatsApp number not registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error:
                  code: not_found
                  message: "No account found with this WhatsApp number."
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [Authentication]
      summary: Reset password with OTP
      description: Reset the account password using the OTP received via WhatsApp.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Password reset successful
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResetPasswordResponse'
        '400':
          description: Invalid OTP or weak password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error:
                  code: invalid_otp
                  message: "OTP is invalid or has expired."
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'

  # ═══════════════════════════════════════════════════════════════
  #  PAYMENTS
  # ═══════════════════════════════════════════════════════════════
  /payments:
    post:
      operationId: createPayment
      tags: [Payments]
      summary: Create a new payment
      description: |
        Create a new crypto payment request. Returns a unique deposit address
        and a hosted checkout URL. The payment expires after `expiry_minutes`
        (default 30 minutes).

        **Flow:**
        1. You create a payment via API
        2. Redirect customer to `checkout_url` or display `deposit_address` and `qr_code_url`
        3. Customer sends crypto
        4. MyCryptoCoin detects the transaction and updates status to `confirming`
        5. After required confirmations, status becomes `confirmed`
        6. Funds are settled to your wallet (`settled`)
        7. You receive webhook notifications at each step
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePaymentRequest'
      responses:
        '201':
          description: Payment created
          headers:
            X-RateLimit-Limit:
              $ref: '#/components/headers/X-RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'

    get:
      operationId: listPayments
      tags: [Payments]
      summary: List all payments
      description: Retrieve a paginated list of payments with optional filters.
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirming, confirmed, settled, expired, failed, refunded]
          description: Filter by payment status
        - name: crypto
          in: query
          schema:
            type: string
          description: Filter by cryptocurrency (e.g., BTC, ETH)
        - name: date_from
          in: query
          schema:
            type: string
            format: date
          description: "Start date (YYYY-MM-DD)"
        - name: date_to
          in: query
          schema:
            type: string
            format: date
          description: "End date (YYYY-MM-DD)"
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: List of payments
          headers:
            X-RateLimit-Limit:
              $ref: '#/components/headers/X-RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
  /payments/{id}:
    get:
      operationId: getPayment
      tags: [Payments]
      summary: Get payment details
      description: Retrieve full details of a specific payment by ID.
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Payment ID (e.g., pay_1a2b3c4d5e6f)
      responses:
        '200':
          description: Payment details
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /payments/{id}/verify:
    post:
      operationId: verifyPayment
      tags: [Payments]
      summary: Verify payment status
      description: |
        Verify the current status of a payment. Use this as a fallback
        if you missed a webhook or want to double-check before fulfilling
        an order. This performs a real-time blockchain check.
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Payment verification result
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentVerifyResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  # ═══════════════════════════════════════════════════════════════
  #  WALLETS
  # ═══════════════════════════════════════════════════════════════
...
  /merchant/api-keys:
    post:
      operationId: createApiKey
      tags: [Merchant]
      summary: Generate a new API key
      description: |
        Generate a new API key for programmatic access. The full key is shown
        only once in the response. Store it securely.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateApiKeyRequest'
      responses:
        '201':
          description: API key created (full key shown only once)
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKey'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

    get:
      operationId: listApiKeys
      tags: [Merchant]
      summary: List API keys
      description: |
        Retrieve all API keys. For security, only the key prefix is shown,
        not the full key.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of API keys
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...